requirements.txt
# N8N Workflows API Dependencies
# Core API Framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.4.0,<3.0.0

# For RAG and Supabase Integration
python-dotenv>=1.0.0,<2.0.0
supabase>=2.0.0,<3.0.0
openai>=1.0.0,<2.0.0